+++ lib/commands/cms.rb
@@ -4,8 +4,11 @@ module Yawast
   module Commands
     class Cms
       def self.process(args, options)
-        uri = Yawast::Commands::Utils.extract_uri(args)
-        Yawast::Scanner::Core.get_cms(uri, options)
+        args.each do |arg|
+          uri = Yawast::Commands::Utils.extract_uri([arg])
+
+          Yawast::Scanner::Core.get_cms(uri, options)
+        end
       end
     end
   end