+++ lib/commands/head.rb
@@ -4,10 +4,12 @@ module Yawast
   module Commands
     class Head
       def self.process(args, options)
-        uri = Yawast::Commands::Utils.extract_uri(args)
+        args.each do |arg|
+          uri = Yawast::Commands::Utils.extract_uri([arg])
 
-        options.head = true
-        Yawast::Scanner::Core.process(uri, options)
+          options.head = true
+          Yawast::Scanner::Core.process(uri, options)
+        end
       end
     end
   end