+++ lib/commands/scan.rb
@@ -4,9 +4,11 @@ module Yawast
   module Commands
     class Scan
       def self.process(args, options)
-        uri = Yawast::Commands::Utils.extract_uri(args)
+        args.each do |arg|
+          uri = Yawast::Commands::Utils.extract_uri([arg])
 
-        Yawast::Scanner::Core.process(uri, options)
+          Yawast::Scanner::Core.process(uri, options)
+        end
       end
     end
   end