+++ lib/commands/ssl.rb
@@ -4,9 +4,11 @@ module Yawast
   module Commands
     class Ssl
       def self.process(args, options)
-        uri = Yawast::Commands::Utils.extract_uri(args)
+        args.each do |arg|
+          uri = Yawast::Commands::Utils.extract_uri([arg])
 
-        Yawast::Scanner::Core.check_ssl(uri, options, nil)
+          Yawast::Scanner::Core.check_ssl(uri, options, nil)
+        end
       end
     end
   end